Add current TransactionController

+++ src/main/java/com/example/personal_budget/dto/CreateTransactionRequest.java
@@ -6,23 +6,22 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 @Data
 public class CreateTransactionRequest {
-  private BigInteger userID;
+    private Long userID;
 
-  @Positive(message = "Amount must be positive")
-  private double amount;
+    @Positive(message = "Amount must be positive")
+    private double amount;
 
-  @NotNull(message = "Type can't be null")
-  private TransactionType type;
+    @NotNull(message = "Type can't be null")
+    private TransactionType type;
 
-  @JsonFormat(pattern = "yyyy-MM-dd")
-  private LocalDate date;
+    @JsonFormat(pattern = "yyyy-MM-dd")
+    private LocalDate date;
 
-  private BigInteger categoryID;
-  private String source;
-  private String description;
+    private Long categoryID;
+    private String source;
+    private String description;
 }
