+++ src/main/java/com/example/personal_budget/dto/TransactionFilterRequest.java
@@ -2,13 +2,12 @@ package com.example.personal_budget.dto;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 @Data
 public class TransactionFilterRequest {
-  private BigInteger userID;
-  private LocalDate startDate;
-  private LocalDate endDate;
-  private BigInteger categoryID;
+    private Long userID;
+    private LocalDate startDate;
+    private LocalDate endDate;
+    private Long categoryID;
 }
